# api/openapi.yaml
openapi: 3.1.0
info:
  title: 127ohoh1 control-plane API (development reference)
  version: 0.1.0
  description: >
    Public reference API. Paths are added as phases land. Payment endpoints are
    simulated, development-only, and fail closed unless explicitly enabled.
    All timestamps are UTC RFC 3339. Errors use the envelope in `Error`.
servers:
  - url: http://localhost:8080
security:
  - session: []
paths:
...
  /v1/endpoints/reserve:
    post:
      summary: Reserve a hostname; returns 402 with an offer if a paid entitlement is missing
      description: >
        Naturally idempotent. An unpaid retry returns the same open offer; a retry
        after settlement (same Idempotency-Key) returns the reserved endpoint.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, plan]
              additionalProperties: false
              properties:
                name: { type: string, description: "Lowercase DNS label, 3-63 chars" }
                plan: { type: string, enum: [tier_1, tier_2, tier_3] }
      responses:
        "201":
          description: Reserved endpoint created
          content: { application/json: { schema: { $ref: '#/components/schemas/Endpoint' } } }
        "200":
          description: Endpoint already reserved by the caller
          content: { application/json: { schema: { $ref: '#/components/schemas/Endpoint' } } }
        "400": { $ref: '#/components/responses/Error' }
        "402":
          description: Payment required (machine-readable offer; see docs/payment-protocol.md)
          headers:
            Cache-Control: { schema: { type: string, enum: [no-store] } }
          content: { application/json: { schema: { $ref: '#/components/schemas/PaymentRequired' } } }
        "409": { $ref: '#/components/responses/Error' }
  /v1/offers:
    post:
      summary: Create (or reuse) an offer without reserving
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, plan]
              additionalProperties: false
              properties:
                name: { type: string }
                plan: { type: string, enum: [tier_1, tier_2, tier_3] }
      responses:
        "201":
          description: Offer
          content: { application/json: { schema: { $ref: '#/components/schemas/Offer' } } }
        "400": { $ref: '#/components/responses/Error' }
        "409": { $ref: '#/components/responses/Error' }
  /v1/offers/{offer_id}:
    parameters:
      - { name: offer_id, in: path, required: true, schema: { type: string } }
    get:
      summary: Offer status with its payment (owner only)
      responses:
        "200":
          description: Offer and payment
          content:
            application/json:
              schema:
                type: object
                properties:
                  offer: { $ref: '#/components/schemas/Offer' }
                  payment: { $ref: '#/components/schemas/Payment' }
        "404": { $ref: '#/components/responses/Error' }
  /v1/payments/{payment_id}:
    parameters:
      - { name: payment_id, in: path, required: true, schema: { type: string } }
    get:
      summary: Payment status (owner only)
      responses:
        "200":
          description: Payment
          content: { application/json: { schema: { $ref: '#/components/schemas/Payment' } } }
        "404": { $ref: '#/components/responses/Error' }
  /v1/payments/{payment_id}/simulate-settlement:
    parameters:
      - { name: payment_id, in: path, required: true, schema: { type: string } }
    post:
      summary: "DEVELOPMENT ONLY: settle a simulated payment"
      description: >
        Fails closed with 404 unless development payments are explicitly enabled
        and app_env is not production. Idempotent: repeated calls grant one entitlement.
      responses:
        "200":
          description: Payment (ENTITLEMENT_GRANTED)
          content: { application/json: { schema: { $ref: '#/components/schemas/Payment' } } }
        "404": { $ref: '#/components/responses/Error' }
...
  /v1/abuse-reports:
    post:
      security: []
      summary: Public abuse-report intake (no account needed; rate limited; nothing is enforced automatically)
      description: >
        Bounded and sanitised. Reported URLs are stored as text and never fetched.
        Detection, triage, case management and appeals are outside the public reference.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostname, category]
              additionalProperties: false
              properties:
                hostname: { type: string, description: "Tunnel host or label being reported" }
                category: { type: string, enum: [malware, phishing, illegal_content, spam, harassment, copyright, other] }
                description: { type: string, maxLength: 2000 }
                url: { type: string, maxLength: 500 }
                contact: { type: string, maxLength: 200, description: "Optional, voluntarily supplied" }
      responses:
        "202": { description: Accepted for human review }
        "400": { $ref: '#/components/responses/Error' }
        "413": { $ref: '#/components/responses/Error' }
        "429": { $ref: '#/components/responses/Error' }
  /v1/admin/abuse-reports:
    get:
      summary: "Operator: list reports (links to the suspension hooks via endpoint_id)"
      security: [ { admin: [] } ]
      parameters:
        - { name: hostname, in: query, schema: { type: string } }
      responses: { "200": { description: Reports }, "401": { $ref: '#/components/responses/Error' } }
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
    edge:
      type: http
      scheme: bearer
      description: Edge service credential (constant-time compared)
    admin:
      type: http
      scheme: bearer
      description: Administrative token; the admin API is absent (404) when none is configured
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 8-128 printable ASCII characters. Reuse with a different request returns 409 idempotency_key_reused.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            details: { type: object }
...
    Endpoint:
      type: object
      properties:
        id: { type: string }
        hostname: { type: string }
        url: { type: string }
        kind: { type: string, enum: [random, reserved] }
        state: { type: string, enum: [PENDING, ACTIVE, SUSPENDED, EXPIRED, RELEASED] }
        plan: { type: string }
        created_at: { type: string, format: date-time }
        last_active_at: { type: string, format: date-time }
...
    Offer:
      type: object
      description: Immutable terms snapshot with an expiry; not a reservation.
      properties:
        id: { type: string }
        plan: { type: string }
        status: { type: string, enum: [open, paid, expired, cancelled] }
        scope:
          type: object
          properties: { type: { type: string, enum: [hostname] }, name: { type: string } }
        price: { $ref: '#/components/schemas/Price' }
        limits:
          type: object
          properties:
            bandwidth_bytes_per_second: { type: integer }
            monthly_transfer_bytes: { type: integer }
        payment_methods: { type: array, items: { $ref: '#/components/schemas/PaymentMethod' } }
        payment_id: { type: string }
        expires_at: { type: string, format: date-time }
        status_url: { type: string }
        checkout_url:
          type: string
          format: uri
          description: "Optional https page where this one offer can be paid in a browser wallet; valid while the offer is open. The link is a capability: anyone holding it can view and pay the offer (never redirect the name)."
    Price:
      type: object
      properties:
        amount: { type: string }
        asset: { type: string }
        interval: { type: string }
        unit: { type: string }
    PaymentMethod:
      type: object
      properties:
        type: { type: string, description: "Network-neutral; the reference build only lists development_crypto" }
        network: { type: string }
        asset: { type: string }
        payment_request: { type: string }
        simulated: { type: boolean }
    PaymentRequired:
      type: object
      properties:
        error: { $ref: '#/components/schemas/Error/properties/error' }
        resource:
          type: object
          properties: { type: { type: string }, name: { type: string } }
        offers: { type: array, items: { $ref: '#/components/schemas/Offer' } }
    Payment:
      type: object
      properties:
        id: { type: string }
        offer_id: { type: string }
        state: { type: string, enum: [CREATED, AWAITING_PAYMENT, SEEN, CONFIRMING, SETTLED, ENTITLEMENT_GRANTED, EXPIRED, FAILED] }
        amount: { type: string }
        asset: { type: string }
        network: { type: string }
        provider: { type: string }
        simulated: { type: boolean }
        failure_reason: { type: string }
        updated_at: { type: string, format: date-time }
...